+++ pilot/pkg/networking/v1alpha3/authentication_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package v2
+package v1alpha3
 
 import (
 	"reflect"
