+++ pilot/pkg/networking/v1alpha3/config.go
@@ -32,6 +32,7 @@ import (
 
 	"time"
 
+	authn "istio.io/api/authentication/v1alpha1"
 	meshconfig "istio.io/api/mesh/v1alpha1"
 	networking "istio.io/api/networking/v1alpha3"
 	"istio.io/istio/pilot/pkg/model"
@@ -101,22 +102,22 @@ func BuildListeners(env model.Environment, node model.Proxy) ([]*xdsapi.Listener
 		// TODO: add listeners for other protocols too
 		return buildGatewayHTTPListeners(env.Mesh, env.IstioConfigStore, node)
 	case model.Ingress:
-		services, err := env.Services()
-		if err != nil {
-			return nil, err
-		}
-		var svc *model.Service
-		for _, s := range services {
-			if strings.HasPrefix(s.Hostname, IstioIngress) {
-				svc = s
-				break
-			}
-		}
-		insts := make([]*model.ServiceInstance, 0, 1)
-		if svc != nil {
-			insts = append(insts, &model.ServiceInstance{Service: svc})
-		}
 		// TODO : Need v1alpha3 equivalent of buildIngressGateway
+		//services, err := env.Services()
+		//if err != nil {
+		//	return nil, err
+		//}
+		//var svc *model.Service
+		//for _, s := range services {
+		//	if strings.HasPrefix(s.Hostname, IstioIngress) {
+		//		svc = s
+		//		break
+		//	}
+		//}
+		//insts := make([]*model.ServiceInstance, 0, 1)
+		//if svc != nil {
+		//	insts = append(insts, &model.ServiceInstance{Service: svc})
+		//}
 		// return buildIngressListeners(env.Mesh, insts, env.ServiceDiscovery, env.IstioConfigStore, node), nil
 	}
 	return nil, nil
@@ -259,8 +260,8 @@ func buildSidecarListenersClusters(
 
 // options required to build an HTTPListener
 type buildHTTPListenerOpts struct { // nolint: maligned
-	config           model.Config
-	env              model.Environment
+	// config           model.Config
+	// env              model.Environment
 	mesh             *meshconfig.MeshConfig
 	proxy            model.Proxy
 	proxyInstances   []*model.ServiceInstance
@@ -350,32 +351,12 @@ func buildHTTPListener(opts buildHTTPListenerOpts) *xdsapi.Listener {
 	}
 }
 
-// consolidateAuthPolicy returns service auth policy, if it's not INHERIT. Else,
-// returns mesh policy.
-func consolidateAuthPolicy(mesh *meshconfig.MeshConfig, serviceAuthPolicy meshconfig.AuthenticationPolicy) meshconfig.AuthenticationPolicy { // nolint
-	if serviceAuthPolicy != meshconfig.AuthenticationPolicy_INHERIT {
-		return serviceAuthPolicy
-	}
-	// TODO: use AuthenticationPolicy for mesh policy and remove this conversion
-	switch mesh.AuthPolicy {
-	case meshconfig.MeshConfig_MUTUAL_TLS:
-		return meshconfig.AuthenticationPolicy_MUTUAL_TLS
-	case meshconfig.MeshConfig_NONE:
-		return meshconfig.AuthenticationPolicy_NONE
-	default:
-		// Never get here, there are no other enum value for mesh.AuthPolicy.
-		panic(fmt.Sprintf("Unknown mesh auth policy: %v\n", mesh.AuthPolicy))
-	}
-}
-
 // mayApplyInboundAuth adds ssl_context to the listener if consolidateAuthPolicy.
-func mayApplyInboundAuth(listener *xdsapi.Listener, mesh *meshconfig.MeshConfig,
-	serviceAuthPolicy meshconfig.AuthenticationPolicy) {
-	// TODO(mostrowski): figure out SSL
-	/*	if consolidateAuthPolicy(mesh, serviceAuthPolicy) == meshconfig.AuthenticationPolicy_MUTUAL_TLS {
-			listener.SSLContext = buildListenerSSLContext(model.AuthCertsPath)
-		}
-	*/
+func mayApplyInboundAuth(listener *xdsapi.Listener, authenticationPolicy *authn.Policy) {
+	if requireTLS(authenticationPolicy) {
+		// TODO(mostrowski): figure out SSL
+		log.Debugf("TODO Apply authN policy %#v for %#v\n", authenticationPolicy, listener)
+	}
 }
 
 // buildTCPListener constructs a listener for the TCP proxy
@@ -798,7 +779,8 @@ func buildInboundListeners(mesh *meshconfig.MeshConfig, node model.Proxy,
 		}
 
 		if l != nil {
-			mayApplyInboundAuth(l, mesh, endpoint.ServicePort.AuthenticationPolicy)
+			authenticationPolicy := getConsolidateAuthenticationPolicy(mesh, config, instance.Service.Hostname, servicePort)
+			mayApplyInboundAuth(l, authenticationPolicy)
 			listeners = append(listeners, l)
 		}
 	}
